chore(nvim): start nvim-ufo if neither LSP nor coc-extension are available
Wait until LSP or coc-extension are ready before starting nvim-ufo.
Otherwise, start nvim-ufo as soon as filetype is detected.

I've seen less issues with nvim-ufo when it is started *after*
LSP/coc-extension are loaded. But in case neither is available, I still
want to start it, this commits handles that.